// order_service/src/dto/orderRequest.dto.ts
export type OrderLineItemType = {
    id: number;
    productId?: number;
    itemName: string;
    qty: number;
    price: string;
    orderId: number;
    createdAt: Date;
    updatedAt: Date;
};

export interface OrderWithLineItems {
    id?: number;
    customerId: number;
    orderNumber: number;
    txnId: string | null;
    amount: string;
    status: string | null;
    orderItems: OrderLineItemType[];
    createdAt?: Date;
    updatedAt?: Date;
}

export interface InProcessOrder {
    id?: number;
    orderNumber: number;
    status: string;
    customerId: number,
    amount: number,
    createdAt?: Date,
    UpdatedAt?: Date,
}